==> notebooks/2.0-dataset-api.ipynb <==
#### Consolidación del dataset para el desarrollo de consultas

- Para desarrollar la API descrita en la propuesta de trabajo (README.md) se hizo necesario para optimizar recursos de computo y memoria consolidar un solo dataset para responder las consultas y los endpoints del proyecto.

- Es por esto que, considerando las columnas que se deben consultar y optimizando el tamaño del dataset para fines demostrativos en este proyecto se consolida este dataset.

- Como lo anuncia la propuesta de trabajo, hemos resumido el análisis de sentimiento en una nueva columna sentiment que contiene si la reseña que tiene el juego es positiva, neutral o negativa. Seguiremos trabajando para los endpoints usando exclusivamente estos juegos con reseñas para que la muestra de nuestro proyecto este orientada a esos juegos que fueron calificados y que nos interesa conocer por sus desarrolladores y las recomendaciones de los usuarios.

- Se harán varios 'merge' de los datasets anteriores para alcanzar lo anteriormente descrito, primero se unirá el dataset de reseñas (users_review_df) con el dataset de usuarios y juegos (users_items_df) y posteriormente se fusionará con el dataset de juegos de steam (steam_games_df). 

import pandas as pd

# Carga de los archivos
file_path_games = '../data/processed/processed_steam_games.csv'
steam_games_df = pd.read_csv(file_path_games)

file_path_reviews = '../data/processed/processed_users_reviews.csv'
users_reviews_df = pd.read_csv(file_path_reviews)

file_path_user_items = '../data/processed/processed_users_items.csv'
users_items_df = pd.read_csv(file_path_user_items)

# Uniendo users_items_df con users_reviews_df
combined_reviews_items_df = pd.merge(users_reviews_df, users_items_df, on=['user_id', 'item_id'], how='inner')

# Cambiando el nombre de la columna para el merge con steam_games_df
steam_games_df.rename(columns={'id': 'item_id'}, inplace=True)

# Combinando con steam_games_df
df_complete = pd.merge(combined_reviews_items_df, steam_games_df, on='item_id', how='left')

# Reemplaza los valores NaN de la columna sentiment con la opción por defecto 1: neutral (si es necesario)
df_complete['sentiment'] = df_complete['sentiment'].fillna(1)

# Mostrar el DataFrame resultante
df_complete

Luego de consolidar nuestro dataset, vamos a eliminar los registros duplicados si existen y hacernos cargos de registros con valores nulos en géneros, año de lanzamiento y horas totales de juego que son columnas necesarias para las consultas.

# Verificamos los valores nulos resultantes
df_complete.isna().sum()

# Manejo de valores NaN
df_complete = df_complete.dropna(subset=['genres', 'release_year', 'playtime_forever'])

# los precios nulos los dejaremos en cero
df_complete['price'] =  df_complete['price'].fillna(0)

# Verificamos valores nulos
df_complete.isna().sum()

# Vemos los valores únicos de la columna 'price' para verificar que quede como variable numérica
df_complete['price'].unique()

# Reemplazamos valores de price que sean cadenas de strings
df_complete['price'] = df_complete['price'].replace({"Free To Play": 0, "Play for Free!": 0, "Free Mod": 0, "Free HITMAN™ Holiday Pack": 0,
                           "Free to Use": 0, "Third-party": 0, "Free to Play": 0, "Free":0})

df_complete['price'].value_counts()

# Eliminación de duplicados
df_complete = df_complete.drop_duplicates()

df_complete

Ya con el dataset depurado, lo guardaremos para que sea el dataset definitivo para las 5 primeras consultas de la API

# Guardamos el dataframe procesado para las consultas de la API
output_file_path = '../src/data/dataset_full.csv'
df_complete.to_csv(output_file_path, index=False)

output_file_path

El desarrollo de las funciones para las consultas se realiza en el archivo funciones.py que posteriormente son usadas en el main.py
